requirements.txt
pymongo==4.5.0
bcrypt==4.0.1
black==23.11.0
asgiref==3.7.2
openai==1.54.3
aioboto3==13.2.0
pillow==10.3.0
quart==0.19.6
quart-cors==0.7.0
Quart-JWT-Extended==0.1.0
motor==3.4.0
hypercorn==0.17.3
httpx==0.27.0
watchtower==3.2.0
aiosmtplib==3.0.2